requirements.txt
pycryptodome>=3.6.4